UPDATE: add routines to enforce particle-hole symmetry in the bath.
UPDATE: following the update in SciFor OPTIMIZE routines with introduction
of two new methods to perform CG minimization, we introduced a flag to
chose which method to use: cg_method= 0:Num.Rec. [default],
1:minimize (f77 routines used in old ED codes), 2: CG+.

tests with attractive model (look to RDMFT) shows that cg_method=1 is indeed
not very powerful and strongly depends on the initial seed.