Version 2.1 better calculation of Kinetic energy is introduced.
The kinetic energy is now evaluated as
<K> = Tr{H(k)*G(k,iw)}

Subtraction and analytical treatment of the tails is performed to
improve the results. Tests show significant improvement versus
previous procedure Tr{\Delta*Gimp}.
In turns previous method was limited to non-hybridizing systems.

The calculation of kinetic energy requires action from the user.
This potentially breaks the reverse communication strategy.
The calculation of the energy proceeds as follow:
local terms are evaluated just after the solution of the impurity
problem using thermal averages. These are stored in global
variables and always wrote to file.
Kinetic energy calculation can be performed on-call using
a generic routine in ED_MAIN. This latter requires:
H(k): the (LDA) matrix hamiltonian.
[W(k): the weights of each k-points, optional]
Sigma: the matrix self-energy of the problem.

This way is safe with respect to problem where the structure of
the lattice problem is somehow different from that of the impurity
problem itself, e.g. independent sites, spin-orbit systems, etc..

If the kinetic energy routine is invoked, kinetic energy
calculation is performed and the result is overwritten to file,
with a new columns dictionary.

- Reshape of the code:
to better hide ED variables and routines to the user we introduced
a new layer for ED_MAIN, which takes the place of DMFT_ED. This
module is private and only some routines are public.
The DMFT_ED module is now a simple wrapper of the necessary ED modules
and procedure that should be visible to the user.