update:
BUG FIXED in energy calculation in the superconducting case

-added routines for internal energy calculations in the superconducting (SC) phase.
Subtraction and analytical summation of the tails performed as in the normal(N)
phase. Tested computing the condensation energy for the bethe and 2d_square lattice.
-added routine enforcing non symmetry breaking calculations with ED_SUPERCOND=T.

role of ED_SUPERCOND flag modified:
now it only decide whether to compute or not
the anomalous parts of the greens function.