bug fix: little bug found in the full diagonalization of H in the parallel mode (i.e. in small sectors). MPI_BCAST was called with wrong dimension of the H matrix (dim instead of dim*dim). fixed.
update:
the finite temperature channel has been updated with a more sophisticated
way to handle the number of states in the spectrum.
A check is performed after diagonalization of the required states (lanc_states_total) to
determine whether the cutoff condition exp(-beta*(E_i-E_gs)) < e_cutoff is satisfied
or not.
If so (FALSE) then more states are required and lanc_states_total is increased at the next iteration.
If not (TRUE) then one has more states than actually required. Action is taken to trim the list for
the next calculation of observables and GF and lanc_states_total is trimmed down.
Some care is still needed in how to perform the trimming of the lanc_states_total.

update2: removed progress bar in both diagonalize_impurity and build_greens_functions.
Print on the screen a lot of information, unless ed_verbose>2