Update: code fixed to compile and run with GNU fortran compiler.
Several issues present for this compiler:
- ED_EIGENSPACE: pointer in the estate type were not set to null()
in their default definition, so they were in an undefined state.
this led es_free_espace to a memory problem because it could not
check if a pointer was associated or not. FIXED
- allocate problem for seed_random in ARPACK module

ADD:
coded the total spin susceptibility for multi-band systems.
some steps to take, getting ready.