VERSION v3.0.1:
update:
the disorder attractive code was updated to the latest modifications introduced
for the nano objects driver. Namely the driver did not compile because of the
missing *iprint input in get_gloc_lattice and the missing *Nsites in the
get_weiss_lattice routines.

- this is a slight update of the code.
A driver for nano objects has been written in collaboration with Angelo Valli.
This required few modifications namely in the ED_WRAP_WEISS module to deal with a
number Nsites < Nlat of independent sites.

In turn we discovered a severe bug in the evaluation of the Weiss Field in the
_MPI_INEQ parallel configuration that went unnoticed before because testing were
done serially.
The lack of initialization of the Weiss_tmp=zero array led to bad mpi_reducing
which ended up in badly oscillating Weiss Field function screwing up the fit.

- the driver for the BHZ edge states was also wrote.