The nonSU2 version of the code is almost entirely written (with G.Sangiovanni) during my stay at Wuerzburg. This version still does not have the implementation of the Delta function calculation and the chi^2 fit part.
I am thinking about the best strategy to implement the spin-flip amplitude W.
This version has separate V and W hoppings but it is non-satisfactory.

So I decided to change the code so to have
V(Nspin,Norb,Nbath) --> V(Nspin,1-2,Norb,Nbath)
where the second index indicate the existence (2) or not (1) of spin-flipping hopping.

Ideally this amplitude could include also the superconducting one in the ed_mode=superc case.

todo:
- I would like to spread the ED_GREEN_FUNCTIONS module and files so to possibly improve the readibililty
and reuse of routines. Additionally one could also try to shrink the routines that apply operators to the
state vectors.
- I would like to simplify the structure of chi2fit module, so to reuse the functions coded in the ED_BATH module.