update: the CDW-HM driver was tested and fixed with K.Kapcia (leading this branch)
we checked that the code gives reasonable results for both the V=0 and the V!=0 case.
CDW-I phase has been found for V=3,Uloc=2
FL-MET phase has been found for V=0,Uloc=2 && V=0.5,Uloc=2

more tests and comparison with previous results may be required.

bugs:
in the process we found a couple of bugs that we fixed but we did not propagate
to the main branches. Namely:
- ED_GET_GLOC needed to get a suffix featuring the ed_file_suffix string in order
to distinguish possibly different inequivalent sites.
- the same routine improperly plotted the functions on the real-axis (Greal) against
the matsubara frequencies (wm) rather than wr (real-freq. array).
- ED_ENERGY we now use print_hloc from ED_AUX_FUNX to plot the local (with respect to routine)
version  of Hloc. In addition we fixed some possible error in the dimension of Wtk array in the
1-b procedure, where it was declared as size(Hk,1) but Hk  only has 1 dimension.