BIG UPDATE. -----------------------WARNING---------------------------------- THIS VERSION OF THE CODE NEEDS A THOROUGH TESTING PHASE ----------------------------------------------------------------
This is a new version of the DMFT-ED code which introduced several changes.

Here is a list:
- ED_BATH:
the user bath (bath_) is now a one dimensional array.
The module now contains in unique procedures the calculation of the delta and G0 (weiss)
functions for a given dmft effective bath (dmft_bath).
An interface allows to use the user bath. Each procedure evaluate delta or weiss function
dependending on the setup of bath_type=normal,hybrid and ed_mode=normal,superc.
The module features also new routines that allows to push/pull part of the bath as
used in the \chi^2 fit (usually denoted a or array_bath) into the dmft_bath. This
step is necessary to use the procedures in ED_BATH inside the evaluation of the
\chi^2 fit. Before the delta or G0 functions or gradient were evaluated inside the
ED_CHI2FIT module independently. I wanted to use the same lines of code.

- ED_VARS_GLOBAL:
Some of the variables setting up the dimensions of the problem have been re-named.
Ntot = Nlevels
NN   = Nhilbert
etc..

- ED_GREENS_FUNCTIONS:
I developed the calculations of the anomalous-mixed green's functions for the
bath_type=hybrid + ed_mode=superc channel.
The latter channel is now almost complete.
WARNING THIS CHANNEL IS NOT TESTED. A stop condition is still present in the
checking part of the init_ed procedure.
The calculation of the Anomalous GF in the superconducting case was updated
with a Lanczos scheme which do not involve knowing the symmetry between the
G_21 and the G_12 components of the Nambu Green's functions. This comes at the
cost of applying more operators to the actual eigenstate |v>.

- ED_CHI2FIT:
This module has been heavily updated. The routines have been symmetrized and
redefined. The \Chi^2 functions passed to the CG minimization now entirely
rely on the delta/G0 weiss functions present in ED_BATH.

- ED_ENERGY:
the routines now behaves as in the ED_WRAP_ENERGY module, namely
procedures are functions which returns an array real(8),dimensioN(2) Eout.

Less important updates have been included in this version.

update 1: fixed the bug (already solved by G.Mazza) of the REV variable
in the make.inc file which prevented from storing the
SH1 - hash of the actual version of the ED code.

update 2: fixed ed_bhz for compilation with this new version

update 3: multi-band multi-spin checked with BHZ model solution.
So far for the repulsive models, results are always identical to
machine precision.

--> THERE APPEARS TO BE A SMALL PROBLEM WITH THE bath_type=HYBRID
--> + cg_scheme=WEISS.
The fit is very slow and the results in the BHZ model look bad.
This problem was not introduced with the new version but existed
also in the master branch. For example the tdd-PAM results for 2bands
were obtained using cg_scheme=delta and worked fine.

update 4:
There is a problem for the attractive model. While the impurity
GF at 1loop are almost identical to the master_branch results
(this means that the new way of evaluating the impurity GF is
working fine although a little bit redundant), the Self-energies,
the G0,F0 functions on THE REAL AXIS are not correct.
As a results the code converges fine and the results for the observables
are identical (to the given precision) to the master_branch, but the
functions on the real axis are wrong. CHECKING IT. 1.May.2015

update 5:
attractive code debugged. Results are now in agreement with those
of the master branch. A couple of minor bugs were present in the
functions evaluating Delta(w) in ed_bath_delta_bath_real.f90.