SECOND BIG UPDATE: ------------------- W A R N I N G -------------------------
--------> THIS CODE HAS NOT BEEN RE-TESTED <---------------

this is an update on top of the previous version featuring a big change with
respect to the master branch.

In this version:
- direct and inverse non-interacting Green's function for the Anderson
model are evaluated inside ED_BATH and used both in the Chi^2 fit module
ED_CHI2FIT and in the GF module ED_GREENS_FUNCTIONS

Further tests are in order now to check that everything is working.
This version is ready to import the nonSU(2) symmetric development
from the devel_nonsu2 branch.