Update:
put this version up to date (manually) with the latest update of the new_chi2fit branch (ahead of the master branch).
In this update we introduced procedures to evaluate the direct and inverse non-interacting Green's functions of the
Anderson model (g0and and invg0), inside ED_BATH.
These routines still lack the extension to the nonsu(2) case and are necessary to get the Self-energies in the
ED_GREENS_FUNCTIONS module as well as to perform the chi^2 fit in ED_CHI2FIT module.

A bientot.