UPDATE:
- the chi^2 fit is updated, with respect to the latest change introduced in this branch.
The goal of previous update was to unify the calculation of Delta and G0 functions based on the
dmft effective bath. This was implemented using a rather large number of procedures in the
ED_BATH module. It turned out that this change introduced a rather big slow down in the fit part.
So I decided to fall back to an updated of the original form:
 . spread of the bath array used in the chi^2 minimization (a)
 . evaluation of the Delta, \grad Delta and G0 functions
are performed directly inside each ed_chi2_xxx file, where the minimization of the \chi^2 function
is done for each bath_type normal/hybrid & ed_mode normal/superc case.

- in the ED_GREENS_FUNCTIONS module the evaluation of the self-energies and the print part have been
separated for better reading and debug.

TODO:
discussing with M.Capone, he proposed the idea of implementing a rather general and more useful form of the
bath, so that the hybridizations are always diagonal while all the complications are included in the energies and
hopping within the bath. This has the main advantage of reducing the suppression of a particular component of the bath
green's functions to the annihilation of a particular hopping, without affecting the other component. An implementation
of this is for example in the superconducting form of the bath where the Delta_k amplitudes are defined in the bath levels
and setting them to zero is enough to suppress anomalous GF.

In this implementation the \Delta function should have the expression:
Delta = V^+ . (z - E)^{-1} . V
with V is the matrix storing the diagonal (at least in some pseudo-blocks) hybridization amplitudes and
the E matrix storing the levels and hopping amplitudes inside the bath. A similar form is obtained for the superconducting case.