DEVEL_UPDATE:
development concluded.
updated code has been tested with success for the three driveres
ed_hm_bethe
ed_bhz
ed_ahm_bethe

finding excellent or very good results with respect to master
branch and previous version of this same branch.

In the process I fixed few bugs in
-ed_chi2_fitgf_normal_normal/superc

and restored the former method for the calculation of the
anomlaous GF in
-ed_greens_funcs_build_gf_superc.f90
which assume a precise symmetry between the G_12 and G_21.

Waiting to test the ed_ahm_stripe code from G.Mazza.
This version will now be merged with master branch.