UPDATE devel_nonSU2:
This version extends the latest version in devel_update - master.
It has all the latest features and debugs introduced so far.

It was tested for:
- ed_hm_bethe
- ed_bhz

with perfect results (numerical precision agreement with former data).

The nonSU2 extension now include a bath with:
- eps: spin dependent energy levels (diagonal)
- vps: same-spin hybridizations (non-diagnoal)
- ups: spin-flip hybridizations (non-diagonal)

which are used throughout the code.
To evaluate the Delta/G0 function we use the W_{ab}^{ss`} matrix
starting from the vps/ups hoppings.

The Delta/G0 are included directly in both the \chi^2 fit part AND
the ED_BATH module (used in ED_GREENS_FUNCTIONS).

More tests (there is still a possible tiny bug in ed_ahm_bethe)
are in order.
The nonSU2 part is still NON-TESTED.

This version is ready to be merged with devel_update branch.

update: added a program driver for F.Petocchi to tackle the
nonSU2 problem in strained STO systems. This is a starting
point for the problem, taken from the ed_bhz and stripping
away unnecessary routines and procedures.