UPDATE DEVEL_UPATE (!):
This version of the code is a merge with devel_nonSU2.
It features the calculation of broken SU(2) symmetry systems.
For more information see past logs.

Tests for the repulsive cases give EXACT results with respect to the
master branch.
Tests for the attractive cases give slightly different results during the
DMFT iterations, caused by a tiny mismatch in the fit procedure and GF
evaluations. The final results are btw in excellent agreement with the past.
[The difference being quite arbitrary].

This problem was investigated thoroughly and shed some light on other minor
bugs.
The problem here is entirely due to the Nambu structure of the Green's functions
which requires many inversions of 2x2 matrices, thus leading to numerical
difference (not necessarily errors, just differences) which propagate in the fit
and the entire DMFT iterations.

------------------------> WARNING <-------------------------
------------------------> WARNING <-------------------------
We declare this code version stable for the SU(2) symmetric part
------------------------> WARNING <-------------------------
------------------------> WARNING <-------------------------

Tests for the nonSU(2) part are in order and will be done in the next
few days.

This is ready to be merged into master branch.