UPDATE:
<ironic mode:ON>
The good and old GNU fortran compiler is maintained by a group of conservative assholes
who thinkgs that in fortran is not useful after all to access arrays with a rank > 7
so they limit themselves to index the ranking with a small bit number 2**3.
Nothing like the bad and ugly people at INTEL who allows to access 31 dimensions and who
are clearly paid by the high-energy string people to make a propagand for the many dimensions,
notwithstanding the money contribution from the DMFT community.
<ironic mode: OFF>

Given that gfortran does not support a rank > 7 we changed the code to get Gij in the
superc channel: ed_get_gij_lattice_superc
By separating the 1:2 normal:anomal components of the G_ij into 2 different arrays:
G(Nlat,Nlat,....,Lfreq)
F(Nlat,Nlat,....,Lfreq)