bug: FP and GS found a serious bug, namely the imp GF does not respect the Dyson equation, so it is not correctly evaluated in the nonSU2 channel. After a first look at the code in the nonsu2 channel nothing suspicious came out.
debug steps: (here we do not make distinction between normal and hybrid bath, but we are consistently using
bath_type=HYBRID)
- solved the 1band Hubbard model with the ed_mode=normal && nonsu2 channel at half-filling and non-interacting limit
noticed that in the normal channel dens=1 and docc=1/4 as expected.
in the nonsu2 channel dens!=1 (0.87) and docc!=1/4 (0.16) this is very bad and traceback the problem
to miscalculation of the Hamiltonian, which can possibly explains the misbehavior of imp GF.

- AA set the spin-flip hopping of the bath in the nonsu2 channel to zero in the ED_BATH/init_ed_bath routine.
This lead to a correct calculation of the Hamiltonian, its ground state and the imp GF as in the ed_mode=normal channel.
[this should be enough to demonstrate that H is to blame for this bug].

most important changes to the code with respect to the latest commit:
/ED_BATH.f90
-    if(ed_mode=="nonsu2")dmft_bath_%u(:,:,:) = dmft_bath_%v(:,:,:)
+    if(ed_mode=="nonsu2")dmft_bath_%u(:,:,:) = 1d-2!dmft_bath_%v(:,:,:) or zero

/ED_GREENS_FUNCTIONS.f90
@@ -753,7 +753,7
                 l=l+1
                 getIorb(l)=iorb
                 getIspin(l)=ispin
-                getJorb(l)=jorb
+                getJorb(l)=iorb
                 getJspin(l)=jspin
this bug associate jorb=0 to getJorb which produced an error (seg. fault) in print_sigma_nonsu2 for bath_type=normal.

/ed_bath_delta_bath_mats.f90
-60,8 +60,8 @@ function delta_bath_mats_main(x,dmft_bath_) result(Delta)
-                 do ih=1,Nspin
-                    do i=1,L
+                 do i=1,L
+                    do ih=1,Nspin
                        Delta(ispin,jspin,iorb,iorb,i) = Delta(ispin,jspin,iorb,iorb,i) + sum( wps(ispin,ih,:)*wps(ih,jspin,:)/(x(i) - hps(ih,:)) )

@@ -114,10 +114,10 @@ function delta_bath_mats_main(x,dmft_bath_) result(Delta)
-                    do ih=1,Nspin
-                       do i=1,L
+                    do i=1,L
+                       do ih=1,Nspin
                           Delta(ispin,jspin,iorb,jorb,i) = Delta(ispin,jspin,iorb,jorb,i) + &
-                            sum( wops(ispin,ih,iorb,:)*wops(ih,jspin,jorb,:)/(x(i) - hps(ih,:)) )
+                               sum( wops(ispin,ih,iorb,:)*wops(ih,jspin,jorb,:)/(x(i) - hps(ih,:)) )

similar changes have been added into /ed_bath_delta_bath_real.f90.
This very serious bug was discovered by F.Petocchi while debugging the nonsu2 channel, was introduced by me
when changing the code of delta_bath to get delta for all matsubara frequencies rather than for a single one.

Other small changes here and there.
Restored the make.inc use in Makefile!