update: some debug lines added to find out the bug that makes the nonSU2 code not working. Ready to merge with previous changes stored in debug_nonsu2_debugAA.
After a long check, we find out that the error is most probably in the determination of the
spectrum. It is still not cleear if the bug lies in the construction of the H or in the
Lanczos procedure itself.

Solving with Lapack diag routines rather than Lanczos seems to fix the problem with the density
at the first loop. But there is still a possibility that the code does not work for additional reasons.
The fit anyway must be changed because it is wrong and surely will make the code not converging.