few steps forwards in the debug process.
The problem with the determination of the low-part of the spectrum or gs it is
intrinsecally related to the equality of the V_sh & V_sf initial hopping parameters.
We still do not understand its origin though, I guess it is a too high level of degeneracy
which makes for Lanczos very hard, too much even for Arpack!!, to properly determine the
groundstate vectors. All the tests performed to the Lanczos process did not point out
significative problems.

The chi^2 fit procedure got some important problems related to the possible symmetry
of up-dw spins. If ed_spin_sym=T condition was required (i.e. spin up & spin dw should
have identical parameters) the bath array *a* used in the chi2 fit routines should
have been extracted in the proper way and it was not.
Now I enforced the fit of only the spin up paramters, which are copied to the spin dw part
at the end of process.

Using the initialization:
V_sh = 0.1
V_sf = V_sh/10
the code converges for U=0 in 3 loops (as expected).

The code also converges for U=1 but this is almost a miracle given that I realized that the
fit has someting wrong in it, i.e. the fit is not good at all. There is a possible factor 2 missing.
More checks to be done, but we are on the right path.