update: further testing show that the code seems to start working now.
Few updates to the code and in the driver ed_hm_bethe that now is happy
both with ed_mode=normal and ed_mode=nonsu2.
Results are in excellent agreement in the two channels for U=1 and initial choice
of V_sf=0 (added an input variabile ed_vsf_ratio to set this).

--> Results for V_sf=0.01 (1/10 of V_sh) looks in agreement too, but there is a
misbehavior of Delta: ImDelta does not go to zero... very suspicious! <---