update:
we keep debugging this weird code. Some things have been changed in this new version:
- added a limit to number of iterations in lanc tri-diagonalization in Gimp.

- the way spin-hold and spin-flip parameters are handled has been slightly updated (after having
experimented with Nspin=1 or Nspin=2). It became clear that the solution of this problem nonSU2
requires Nspin to be 2, thus we introduced a parameter ed_para to enforce the existence of a paramagnetic solution.

- There was an error in the evaluation of the Delta function (Whel(jspin,jh,:)---> Whel(jh,jspin,:))
in ed_bath_delta_bath_mats/real.f90

- We discovered the origin of the misbehavior at the first loop originating from the condition V_sh=V_sf.
This led to a degeneracy of the ground state, that if not resolved  by the Arpack-Lanczos caused an avalanche
effect on the evaluation of the Gimp and observables.
 This is now fixed (hopefully) by setting the initial number of states per sector required to Lanczos to
lanc_nstates_sector > 1 (e.g. 6).
To control the computational cost we introduced other variable to control the size of the block used in Lanczos:
lanc_ncv_factor and lanc_ncv_add (so that NCV  = Nblock = lanc_ncv_factor * lanc_nstates_sector + lanc_ncv_add).

----> the code is still not working, i.e. does not converge. Larger sizes have to be worked out but there is
possibly another (the last?) bug to fix <-----