v4.0.3
- merged the +devel_chi branch into the master (stable) one.
The calculation of the spin suscptibility was improved and tested.
The calculation of the charge susceptibility is now added to the code.
The corresponding routines are included in ed_greens_funx_build_chi_dens.f90, which
is included in the ED_GREENS_FUNCTIONS.f90 module. The calculation of the charge
susceptibility is performed is input variable chiflag=T.

The analytic expressions of the susceptibility spectral sum, used to
accumulate the function in the Lanczos procedure, are the same for spin and charge
(and for any susceptibility chi_A for an operator A).
This means that if chi_spin is correct (as we tested) also chi_dens should be correct.

In any case, further tests are in order to exclude any possible bug.

Files involved in this update:
ED_GREENS_FUNCTIONS.f90
ed_greens_funx_build_chi_dens.f90

pour tous les reinsegnements: adriano.amaricci@gmail.com