v4.1.0: in this version all the _WRAP_ modules for the inequivalent sites problem have been merged in their respective single-site modules (e.g. ED_WRAP_GLOC ---> ED_GLOC).
- ED_GLOC.f90: all the routines are independent on the input Nlat,Nspin,Norb values,
this required to introduce some changes and numerous checks. The latter are performed
using the newly created assert_shape functions in SCIFOR (SF_MISC). All the parameters
are reconstructed from the size of the input arrays. In this way the self-consistency
part can be performed in slightly more independent way from the impurity.
The outcome is twofold:
 + one can easily extend the calculations to p-d problems (two sets of electrons, one
interacting, the other not).
 + one can extract the get_gloc  part from this code and put in the DMFT_TOOLS library.
This is reasonable because the problem of getting gloc is a general one, independent
of the solver and perfectly defined once the H(k) and the Sigma functions are specified.

- ED_WEISS.f90: changes similar to ED_GLOC.f90. Some semplifications are also introduced
in the calculations of the weiss field for the ineq. sites case.

- ED_ENERGY.f90: this module was already quite independent of the input paramters.