This development branch was used to update the Long Range Ordered Magnetic solution of the 2d and 3d BHZ model.
With the occasion AA updated the ED_BATH module in the code.

- the ED_BATH module has been splitted into three sub modules:

 + ED_BATH_DMFT : which contains the procedures related to the
data type type(effective_bath) which conventiently stores the bath
in the ED code. These procedures are used solely inside the code and
are opaque to the user.

 + ED_BATH_USER : which contains the procedures related to the
user defined bath, i.e. the rank 1 array defined in the driver or
calling code by the user. I included new procedures to perform
operation on single component of the bath directly from the
driver.
The problem is that the user bath stores the bath parameters in a
slightly complicated way (all in a single column). So if the user
needs to perform a simple operation on a particular component, say
the energies \epsilon(i=1,Nbath), of the bath he may have hard time
retrieving the correct components. The new procedures named:
get_component_bath
set_component_bath
copy_component_bath
allow to overcome this problem, reading, writing or copying specific
components. For more information read ED_BATH_USER.

 + ED_BATH_FUNCTIONS: which contains the procedures related to the
evaluation of the bath functions: \Delta, G0 and G0^-1.
>> ACTHUNG: these procedures were in specific files up to now. The
files are not used anymore, so I moved them into the hidden .repo
directory. GIT mark them deleted.
Keep this in mind when merging new branches, i.e. debug_nonSU2. <<<

The BHZ AFM 2d/3d codes are updated or written and tested.
Ready for the production cycle.

todo:
- create routines to save and read local functions such as Gloc, Sigma
or other, using the print_gloc template in ED_GLOC.f90