v4.2.5:
small updates from the debug of the ED_AHM_DISORDER driver.

- ED_BATH_DMFT:
the scratch choice of the V parameters in the bath, from the nonSU2 branch
is extraordinarily bad for the superc channel, so I restored the old proposition:
v=max(0.1,1/sqrt(Nbath)). Actually there should be a sum rule to enforce here so that would
be the best choice.

- ED_DIAG:
fixed a bad behavior of the eigenvalues_list printing.

- ED_GLOC:
small checks. nothing important.

- Makefile/make.inc:
restored the working version (compromised by F.Petocchi during the hard nonSU2 debug).

- ed_ahm_disorder/square: small updates to debug the codes.
>> The results of the debug are in test/TEST_AHM_DISORDER. <<
The possible misbehavior is associated to the missing specification of the ts_hopping parameter
in the creation of the 2d_square tight binding Hamiltonian (sitll using the old routine).
Anyway the attractive disorder code and the square bulk both work fine.
single-site has excellent agreement with the old code and the Bethe lattice results.
lattice code has perfect agreement with the bulk single-site code at one loop. At all
loop/convergence the results are affected by numerical errors from the fit. So some sites
have the wrong density. In any case the resulting functions and observables are in very good
agreement.

I can not trace back what was wrong with the original version of this code in the RDMFT+ED
version. For the time being it is better to remove or redo the initial set of disorder
realization.