NON-CRITICAL update v4.2 (mostly nonSU2 channel) - fixed a small bug in the pre-compilation flag of the Sr2IrO4 which associated the MPI universe to the Lanczos set rather than the MPI_INEQ set. Now the code runs but it is quite slow and we are waiting for it to end the first MPI run.
notes:
+symmetrization of the spin orientatin is probably mandatory both for stability of the
convergence and to speed-up the runs. At the moment this is possible only in the
fit procedure (ED_PARA=T enable fit of the up spins only, maybe we can solve somehow
only the up GF)

+there is a possible issue (non-critical) with the too small choice of the initial Vps parameters
of the bath, those generaterd from scratch. Now these are about 0.01 in the nonSU2 channel.
It would be useful to start from a less atomic-like initial condition and possibly add
some noise to the extracted Vps.