UPDATE:
The ed_get_gloc_lattice routine, extended to include the tridiagonal matrices,
has been tested for the BHZ-2D Edge code against the working Tight Binding code
using the same algorithm. This ensure that the routine it is working.

The code has been then tested at 1st loop comparign the direct inversion against
the tridiagonal iterative inversion. The results are correct and identical within
1.d-15.

Speed-up of the performances is rather large, both in terms of memory and cpu-time
for the inversion. We tested up to 40 layers in the ED code, but up to 100 layers
in the Tight Binding code (using the same algorithm).