update:
Here is a clean, elegant a simple way to implement the necessary changes to
generalize the MPI implementation. The goal is to extend the interface procedures
(some of those contained in ED_MAIN) the MPI communicator as the first input variable.
This communicator is defined locally (in the procedure) and then, when necessary, exported
to the global variable ED_MPI_COMM communicator.

While this is easy to do, in practice it requires to define each of these procedures differently
whether or not one is compiling parallel. Compiling just in parallel (dropping serial) is not
a good idea as this will require the user to have MPI even to run serial.

One way to perform this conditional compilation can be achieved using #define Directive within
precompilation. This enable to define the set of input variables of a procedure depending on the
definition of the macro _MPI. Here is an example:

>>>>>>>>>>> TEST PROGRAM <<<<<<<<<<
!compile with:
!gfortran -cpp -D_MPI (or -D_noD or nothing if not)
!ifort -fpp -D_MPI

program testPP
  implicit none
  print*,"with MPI"
  call foo(1,2,1d0)
  print*,"no MPI"
  call foo(2,1d0)

contains

!here MPI_IN will be substituted with that specified above, so including or not the MPI_COMM variable :-)
  subroutine foo(MPI_IN)
    integer :: MPI_COMM
    integer :: int
    real(8) :: double
    print*,MPI_COMM
    print*,int
    print*,double
  end subroutine foo
end program testPP

the produces:
with MPI
           1
           2
   1.0000000000000000

no MPI
           2
   1.0000000000000000