BUG FIX + NEW FEATURE
This version enable to exactly restart a finite temperature calculation
from a previous result. This require to read the states list (conventiently
store in a file to be renamed .restart) and reshape the histogram of the
states evaluated per sector.

In doing this we (AA and AV) found two bugs that we fixed:
 >> 1. the file neigen_sector, storing the information of how many eigensolutions
per sector have to be evaluated, WAS NOT SAVED from one independent site to the other.
Finite temperature calculations of inhomogeneous systems could be affected by this bug.
The bug is fixed by saving the neigen_sector vector and lanc_nstates_total per site to a
specific array neigen_sectorii and lanc_nstates_totalii in ED_MAIN. At each call of ed_diag
in the ed_solve_lattice procedure, the corresponding vector and lanc_nstates_total is retrieved.

 >> 2. the addition and removal of states in the state list was plagued by a bug concerning the
twin states. To cut a long story short: the add procedure included a possible twin state, so that
rather than one, two states are added to the list, while the popping procedure removed only
the last state, with higher energy, irrespective of the fact that this was a twin state or not.
This left a twinless orphaned states in the list at energy higher than the cutoff.
We fixed this by modifying the es_pop_state in ED_EIGENSPACE to remove a state and its twin.
As well we modified the analysis procedure that fixes the state_list at each run (trim or extend).

TEST: we tested the updated code with the ed_nano calculations of Angelo (namely the
only ones at finite temperature and inhomogeneous setup). At one loop the code
reproduce almost exactly the results of previous calculations (irrespective of the actual
convergence of the calculation that we will check).