v4.3 BUG FIX + UPDATE
This update fixes a serious bug in the calculation of the Kinetic energy (Ekin)
in the superconducting channel (ed_mode=superc).
The previous procedures in ED_ENERGY.f90, both for the single-site and the
RDMFT case (_lattice), subtracted the repulsive tail up to (iw_n)**2 to the
11 component of the Nambu GF which could lead in some cases to a value of
Ekin not enough precise.

This new version evaluates the Matsubara sums in the Ekin calculation
subtracting a BCS tail, obtained from the large iw_n limit of the
known self-energies (normal and anomalous). This corresponds to a
correction of order 1/iw_n.

The BCS contribution is evaluated semi-analytically in the general case and
added back.

This implementation has been tested against an old but well tested algorithm
used also in the IPT code for the single-site case (bulk).
Benchmark against this reference  value are perfect up to machine precision for the
bulk case.
The RDMFT (ineq) algorithm gives a very good benchmark (up to 10^-4) with respect
to the bulk:

U=-2.0 single band Attractive Hubbard model on the square lattice
at half-filling, T=1/400, t_hopping=1/4, Lmats=1000, Nbath=4:
   -0.233277623    0.000000000 <== NEW ALGORITHM BULK
 -0.233277623325074            <== OLD ALGORITHM (adapted from Bethe) BULK
   -0.233208899    0.000000000 <== NEW ALGORITHM INEQ

Further testing of the single-band can be found in the ed_ahm_square driver.

-> I also take the chance to update the setup of ed_file_suffix in the RDMFT
case using the newly introduced zero padding in txtfy (requires latest SCIFOR).
I also fixed the misbehavior in the code related to the absence of reset for
ed_file_suffix after any loop over the inequivalent sites, resulting in a
bad naming convention of some files appearing with the _site<latest_site_index> suffix.