v4.3.1 BUG FIX:
The "2Nlso bug" has been fixed. Wait for confirmation from G.Mazza using Gfortran.

In the previous (bugged) version there was a wrong declaration of matrices involved
in a mat-mat product as below:
--- a/ED_ENERGY.f90
+++ b/ED_ENERGY.f90
@@ -853,7 +853,7 @@ contains
     complex(8),dimension(size(Hk,1),size(Hk,2))                     :: Ak,Bk,Ck,Hloc,Hloc_tmp
-    complex(8),dimension(2*size(Hk,1),2*size(Hk,2))                 :: Gknambu,HkNambu,HlocNambu,Evec
+    complex(8),dimension(2*size(Hk,1),2*size(Hk,2))                 :: Gknambu,HkNambu,HlocNambu,Evec,AkNambu
     real(8),dimension(2*size(Hk,1))                                 :: Eval,Coef,Nk_HF,NkDiag
     !
@@ -975,12 +975,13 @@ contains
        Evec(Nlso+1:2*Nlso,1:Nlso)             =             +  blocks_to_matrix(Self_HF)
        Evec(Nlso+1:2*Nlso,Nlso+1:2*Nlso)      = -Hk(:,:,ik) -  blocks_to_matrix(Sigma_HF)
        call eigh(Evec,Eval)
-       NkDiag = fermi(Eval,beta)
+       NkDiag  = fermi(Eval,beta)
        GkNambu = matmul(Evec,matmul(diag(NkDiag),conjg(transpose(Evec))))

-       Ak = matmul(HkNambu, GkNambu)
-       H0free = H0free + Wtk(ik)*trace_matrix(Ak,Nlso)
-       Bk = matmul(HlocNambu, GkNambu)
-       Hlfree = Hlfree + Wtk(ik)*trace_matrix(Bk,Nlso)

+       AkNambu = matmul(HkNambu  , GkNambu)
+       H0free  = H0free + Wtk(ik)*trace_matrix( AkNambu(1:Nlso,1:Nlso) , Nlso)
+       AkNambu = matmul(HlocNambu, GkNambu)
+       Hlfree  = Hlfree + Wtk(ik)*trace_matrix( AkNambu(1:Nlso,1:Nlso) , Nlso)

i.e. the result of a (2*Nlso)**2  matmul product was stored into a Nlso**2 matrix, raising a possible
segmentation fault error (not noticed with Ifort, but present with GNU fortran).

Fortunately, or not, only the first Nlso**2 block is needed for the trace calculation (corresponding to summing
over the <C^+C> part of the density matrix) and Ifort stored precisely this block into Ak and Bk, incredibly
leading to an exact evaluation of the kinetic energy as compared to the BCS calculation (for the bulk).

The (expected) solution is simply to store the calculation into a larger array (here AkNambu) and then take the trace
over the first Nlso**2 block correctly.
The results of a simple test for the AHM_DISORDER driver for U=2 with Wdis=0 gives:
[16:37 ~/.../test_ed/TEST_AHM_KINETIC_3Mar16]$ cat BULK/U2.0/kinetic_* INHOMO_N10/U2.0/kinetic_*
 -0.233307782    0.000000000   <- BULK ED_ENERGY Nk=10**2
 -0.233307781717670            <- DIRECT CALCULATION (using BCS formulae)
 -0.233208899    0.000000000   <- INHOMOGENEOUS Nlat=10**2